Add proxy to access .m3u8 playlist data from player component
The playlist data resource does not allow CORS, so I need to use a proxy to access the data. I could fetch it from the backend and give the information to the frontend, but because of how these playlists work, the source link provided is important to build the individual segments. I would have to create my own hls player to make this work without a proxy or somehow make the player accept the data from afreeca's servers.