Automate message creation & fix encoding issue
Create functions for automating the creation of messages that are accepted by the server. Not necessarily a clean solution, but since  I have no access to any documentation, it will do for now.

Sending messages using special characters was impossible, due to the message length being inaccurate. This is fixed by converting the string into a UTF 8 array buffer.